TestProjectContext currently scans org.springframework.sbm and the test configuration was picked up by other tests too
